separate all queries into one file each
