+++ backend/src/graphql/queries/changeGroupMemberRoleMutation.ts
@@ -1,0 +1,14 @@
+import gql from 'graphql-tag'
+
+export const changeGroupMemberRoleMutation = () => {
+  return gql`
+    mutation ($groupId: ID!, $userId: ID!, $roleInGroup: GroupMemberRole!) {
+      ChangeGroupMemberRole(groupId: $groupId, userId: $userId, roleInGroup: $roleInGroup) {
+        id
+        name
+        slug
+        myRoleInGroup
+      }
+    }
+  `
+}
