+++ backend/src/graphql/queries/createMessageMutation.ts
@@ -1,0 +1,19 @@
+import gql from 'graphql-tag'
+
+export const createMessageMutation = () => {
+  return gql`
+    mutation ($roomId: ID!, $content: String!) {
+      CreateMessage(roomId: $roomId, content: $content) {
+        id
+        content
+        senderId
+        username
+        avatar
+        date
+        saved
+        distributed
+        seen
+      }
+    }
+  `
+}
