+++ backend/src/graphql/queries/groupMembersQuery.ts
@@ -1,0 +1,14 @@
+import gql from 'graphql-tag'
+
+export const groupMembersQuery = () => {
+  return gql`
+    query ($id: ID!) {
+      GroupMembers(id: $id) {
+        id
+        name
+        slug
+        myRoleInGroup
+      }
+    }
+  `
+}
