+++ backend/src/graphql/queries/joinGroupMutation.ts
@@ -1,0 +1,14 @@
+import gql from 'graphql-tag'
+
+export const joinGroupMutation = () => {
+  return gql`
+    mutation ($groupId: ID!, $userId: ID!) {
+      JoinGroup(groupId: $groupId, userId: $userId) {
+        id
+        name
+        slug
+        myRoleInGroup
+      }
+    }
+  `
+}
