+++ backend/src/graphql/queries/markMessagesAsSeen.ts
@@ -1,0 +1,9 @@
+import gql from 'graphql-tag'
+
+export const markMessagesAsSeen = () => {
+  return gql`
+    mutation ($messageIds: [String!]) {
+      MarkMessagesAsSeen(messageIds: $messageIds)
+    }
+  `
+}
