+++ backend/src/graphql/queries/removeUserFromGroupMutation.ts
@@ -1,0 +1,14 @@
+import gql from 'graphql-tag'
+
+export const removeUserFromGroupMutation = () => {
+  return gql`
+    mutation ($groupId: ID!, $userId: ID!) {
+      RemoveUserFromGroup(groupId: $groupId, userId: $userId) {
+        id
+        name
+        slug
+        myRoleInGroup
+      }
+    }
+  `
+}
