+++ backend/src/graphql/queries/unreadRoomsQuery.ts
@@ -1,0 +1,9 @@
+import gql from 'graphql-tag'
+
+export const unreadRoomsQuery = () => {
+  return gql`
+    query {
+      UnreadRooms
+    }
+  `
+}
