+++ backend/src/schema/resolvers/groups.spec.ts
@@ -3,16 +3,14 @@ import { createTestClient } from 'apollo-server-testing'
 import CONFIG from '@config/index'
 import Factory, { cleanDatabase } from '@db/factories'
 import { getNeode, getDriver } from '@db/neo4j'
-import {
-  createGroupMutation,
-  updateGroupMutation,
-  joinGroupMutation,
-  leaveGroupMutation,
-  changeGroupMemberRoleMutation,
-  removeUserFromGroupMutation,
-  groupMembersQuery,
-  groupQuery,
-} from '@graphql/groups'
+import { changeGroupMemberRoleMutation } from '@graphql/queries/changeGroupMemberRoleMutation'
+import { createGroupMutation } from '@graphql/queries/createGroupMutation'
+import { groupMembersQuery } from '@graphql/queries/groupMembersQuery'
+import { groupQuery } from '@graphql/queries/groupQuery'
+import { joinGroupMutation } from '@graphql/queries/joinGroupMutation'
+import { leaveGroupMutation } from '@graphql/queries/leaveGroupMutation'
+import { removeUserFromGroupMutation } from '@graphql/queries/removeUserFromGroupMutation'
+import { updateGroupMutation } from '@graphql/queries/updateGroupMutation'
 import createServer from '@src/server'
 
 const driver = getDriver()
