+++ backend/src/schema/resolvers/observePosts.spec.ts
@@ -4,7 +4,7 @@ import gql from 'graphql-tag'
 import CONFIG from '@config/index'
 import Factory, { cleanDatabase } from '@db/factories'
 import { getNeode, getDriver } from '@db/neo4j'
-import { createPostMutation } from '@graphql/posts'
+import { createPostMutation } from '@graphql/queries/createPostMutation'
 import createServer from '@src/server'
 
 CONFIG.CATEGORIES_ACTIVE = false
